--- src/main/java/com/umc/networkingService/domain/project/service/ProjectHeartService.java
@@ -1,11 +1,0 @@
-package com.umc.networkingService.domain.project.service;
-
-import com.umc.networkingService.domain.member.entity.Member;
-import com.umc.networkingService.domain.project.dto.response.ProjectLikeResponse;
-
-import java.util.UUID;
-
-public interface ProjectHeartService {
-    ProjectLikeResponse likeProject(Member member, UUID projectId);
-    boolean isLikeProject(Member member, UUID projectId);
-}
